backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
gspread==5.12.0
oauth2client==4.1.3
openpyxl==3.1.2
pandas==2.1.3
python-dateutil==2.8.2
pydantic==2.5.0
fastapi-mail==1.4.1
jinja2==3.1.2
aiofiles==23.2.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0